Updated with modifications to decay, TFs, and policy

+++ Sentegrity/Core Detection/Assertion Storage/Sentegrity_Stored_Assertion.h
@@ -19,7 +19,7 @@
 // date and time first created
 @property (nonatomic,retain) NSNumber *created;
 // date and time first created
-@property (nonatomic) float decayMetric;
+@property (nonatomic) double decayMetric;
